requirements.txt
mcp[cli]>=1.5.0
oracledb>=3.0.0
python-dotenv>=1.0.1
uv>=0.2.0
